refactor(engine): streamline ACP plugin parsing
